app/requirements.txt
aiohttp==3.8.4
aiopyston==1.2.1
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
charset-normalizer==3.1.0
click==8.1.3
fastapi==0.95.0
frozenlist==1.3.3
h11==0.14.0
idna==3.4
multidict==6.0.4
pydantic==1.10.6
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.21.1
yarl==1.8.2
